+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Ace.dart Changes
 
-## 0.5.6-dev+20.9.2014
+## 0.5.6+20.9.2014
 
 - Added the `lineHeight` getter and the `textToScreenCoordinates` method to 
 `VirtualRenderer`.